Move Payload.Node to separated file
